+++ domains/builder/components/about-me/useAboutMe.ts
@@ -1,4 +1,4 @@
-import { useCVStore } from '~/domains/builder/store/builder.store'
+import { useBuilderStore } from '~/domains/builder/store/builder.store'
 import { ref } from 'vue'
 import type { IAboutMe } from '~/domains/builder/shared'
 import type { IFileImage } from '~/types/global'
@@ -12,10 +12,10 @@ const DEFAULT_ABOUT_ME: IAboutMe = {
 }
 
 export const useAboutMe = () => {
-  const { updateSection, curriculumVitae } = useCVStore()
+  const { updateSection, builderState } = useBuilderStore()
 
   const aboutMe = computed({
-    get: () => curriculumVitae.aboutMe || { ...DEFAULT_ABOUT_ME },
+    get: () => builderState.aboutMe || { ...DEFAULT_ABOUT_ME },
     set: (value) => {
       updateSection('aboutMe', value)
     }